phase2_new/backend/requirements.txt
# SalaatFlow Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.128.0
uvicorn>=0.40.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Configuration
pydantic-settings>=2.12.0
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.9

# Phase III - AI Chatbot Dependencies
openai>=1.12.0
python-dateutil>=2.8.2
requests>=2.31.0
mcp>=1.25.0
